// packages/database/prisma-generated/zod/inputTypeSchemas/RegistryUserJoinUncheckedCreateInputSchema.ts
import type { Prisma } from '../../client';

import { z } from 'zod';
import { RegistryUserRoleSchema } from './RegistryUserRoleSchema';

export const RegistryUserJoinUncheckedCreateInputSchema: z.ZodType<Prisma.RegistryUserJoinUncheckedCreateInput> =
  z
    .object({
      id: z.string().cuid().optional(),
      createdAt: z.coerce.date().optional(),
      updatedAt: z.coerce.date().optional(),
      role: z.lazy(() => RegistryUserRoleSchema).optional(),
      userId: z.string(),
      registryId: z.string(),
    })
    .strict();

export default RegistryUserJoinUncheckedCreateInputSchema;